+++ src/main/scala/edu/ie3/simona/agent/grid/DBFSAlgorithm.scala
@@ -121,7 +121,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
         case receivedPowers: ReceivedPowerValues =>
           /* Can be a message from an asset or a message from an inferior grid */
           gridAgentBaseData.updateWithReceivedPowerValues(receivedPowers)
-        case receivedSlacks: ReceivedSlackValues =>
+        case receivedSlacks: ReceivedSlackVoltageValues =>
           gridAgentBaseData.updateWithReceivedSlackVoltages(receivedSlacks)
         case unknownReceivedValues =>
           throw new DBFSAlgorithmException(
@@ -297,7 +297,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
         )
       )
       gridAgentBaseData.gridEnv.subgridGateToActorRef
-        .map { case (subGridGate, _) => subGridGate.getSuperiorNode }
+        .map { case (subGridGate, _) => subGridGate.superiorNode }
         .find(_.getUuid == firstRequestedNodeUuid)
         .map(_.getSubnet) match {
         case Some(requestingSubgridNumber) =>
@@ -615,7 +615,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
     // this means we requested an update of the slack voltage values, but for now don't request (and hence don't expect)
     // updated power values for our power flow calculations
     case Event(
-          receivedSlackValues: ReceivedSlackValues,
+          receivedSlackValues: ReceivedSlackVoltageValues,
           gridAgentBaseData: GridAgentBaseData
         ) =>
       log.debug(
@@ -1159,7 +1159,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
             .map { inferiorGridGate =>
               subGridGateToActorRef(
                 inferiorGridGate
-              ) -> inferiorGridGate.getSuperiorNode.getUuid
+              ) -> inferiorGridGate.superiorNode.getUuid
             }
             .groupMap {
               // Group the gates by target actor, so that only one request is sent per grid agent
@@ -1205,7 +1205,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
       subGridGateToActorRef: Map[SubGridGate, ActorRef],
       superiorGridGates: Vector[SubGridGate],
       askTimeout: Duration
-  ): Option[Future[ReceivedSlackValues]] = {
+  ): Option[Future[ReceivedSlackVoltageValues]] = {
     implicit val timeout: AkkaTimeout = AkkaTimeout.create(askTimeout)
     log.debug(
       s"asking superior grids for slack voltage values: {}",
@@ -1220,14 +1220,14 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
             .map { case (superiorGridAgent, gridGates) =>
               (superiorGridAgent ? RequestSlackVoltageMessage(
                 currentSweepNo,
-                gridGates.map(_.getSuperiorNode.getUuid)
+                gridGates.map(_.superiorNode.getUuid)
               )).map { case providedSlackValues: ProvideSlackVoltageMessage =>
                 (superiorGridAgent, providedSlackValues)
               }
             }
             .toVector
         )
-        .map(ReceivedSlackValues)
+        .map(ReceivedSlackVoltageValues)
         .pipeTo(self)
     }
   }