+++ src/main/scala/edu/ie3/util/scala/quantities/QuantityUtil.scala
@@ -8,14 +8,12 @@ package edu.ie3.util.scala.quantities
 
 import edu.ie3.simona.exceptions.QuantityException
 import edu.ie3.util.quantities.{QuantityUtil => PSQuantityUtil}
-
-import javax.measure
-import javax.measure.Quantity
 import tech.units.indriya.ComparableQuantity
 import tech.units.indriya.function.Calculus
 import tech.units.indriya.quantity.Quantities
 import tech.units.indriya.unit.Units
 
+import javax.measure.Quantity
 import scala.collection.mutable
 import scala.util.{Failure, Try}
 
@@ -266,44 +264,4 @@ object QuantityUtil {
     }
   }
 
-  implicit class RichQuantity[Q <: Quantity[Q]](
-      private val q: ComparableQuantity[Q]
-  ) extends AnyVal {
-
-    /** Returns the smaller of two Quantities
-      *
-      * @param other
-      *   the other Quantity
-      * @return
-      *   the smaller of both Quantities
-      */
-    def min(other: ComparableQuantity[Q]): ComparableQuantity[Q] = {
-      if (q.isLessThanOrEqualTo(other)) q else other
-    }
-
-    /** Returns the bigger of two Quantities
-      *
-      * @param other
-      *   the other Quantity
-      * @return
-      *   the bigger of both Quantities
-      */
-    def max(other: ComparableQuantity[Q]): ComparableQuantity[Q] = {
-      if (q.isGreaterThan(other)) q else other
-    }
-  }
-
-  implicit class RichUnit[Q <: Quantity[Q]](
-      private val unit: measure.Unit[Q]
-  ) extends AnyVal {
-
-    /** Transform some power unit to given unit with the same prefix
-      * @param targetUnit
-      *   the target system unit
-      * @return
-      *   this unit converted to given
-      */
-    def toEquivalentIn(targetUnit: measure.Unit[Q]): measure.Unit[Q] =
-      targetUnit.transform(unit.getConverterTo(unit.getSystemUnit))
-  }
 }